Refactored ClassCompiler. Moved code which compiles method bodies to MethodCompiler. Refactored AppCompiler. Moved code which links the excutable to Linker. Removed class loader functions. Class info is now generated as data. Only classes reachable from a set of root classes will be linked by default. Changed a lot of char* to const char*.

+++ vm/rt/src/luni-kernel/org_nullvm_rt_VM.c
@@ -50,12 +50,12 @@ ClassLoader* Java_org_nullvm_rt_VM_getBootClassLoader(Env* env, Class* c) {
 }
 
 jlong Java_org_nullvm_rt_VM_allocateMemory(Env* env, Class* c, jint size) {
-    return nvmAllocateMemory(env, size);
+    return (jlong) nvmAllocateMemory(env, size);
 }
 
 jlong Java_org_nullvm_rt_VM_getCallbackMethodImpl(Env* env, Class* c, Object* methodObject) {
     Method* method = getMethodFromMethodObject(env, methodObject);
-    return ((CallbackMethod*) method)->callbackImpl;
+    return (jlong) ((CallbackMethod*) method)->callbackImpl;
 }
 
 jlong Java_org_nullvm_rt_VM_getObjectAddress(Env* env, Class* c, Object* object) {
